pyproject.toml
[tool.poetry]
name = "bot"
version = "1.0.0"
description = ""
authors = ["yakhovets-o <yakhovetso@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.3.0"
sqlalchemy = "^2.0.27"
aiosqlite = "^0.19.0"
python-dotenv = "^1.0.1"
redis = "^5.0.1"
aioredis = "^2.0.1"
fake-useragent = "^1.4.0"
arq = "^0.25.0"
alembic = "^1.13.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
isort = "^5.13.2"
black = "^24.2.0"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120
lines_after_imports = 2
multi_line_output = 3
skip_gitignore = true
include_trailing_comma = true
skip = "__init__.py"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



